requirements.txt

fastapi
uvicorn[standard]
python-multipart
torch
torchaudio
agno
chromadb
ollama
transformers
sentence-transformers
python-docx
PyPDF2
pypdf